(chore) Rename meals routes file to be consistent
